+++ CHANGES.rst
@@ -3,6 +3,7 @@ Changes in version 0.9.3 (beta)
 
 * Support using ``TranslatedFieldsModel`` with abstract models.
 * Added ``parler.appsettings.add_default_language_settings()`` function.
+* Added ``TranslatableManager.queryset_class`` attribute to easily customize the queryset class.
 
 
 Changes in version 0.9.2 (beta)